# Speed up docker image building and switch base image to alpine