--- a/.dockerignore
+++ b/.dockerignore
@@ -1,5 +1,5 @@
 #
-#i Licensed to the Apache Software Foundation (ASF) under one
+# Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
 # regarding copyright ownership.  The ASF licenses this file