--- a/.github/workflows/distribution-checks.yml
+++ b/.github/workflows/distribution-checks.yml
@@ -22,6 +22,7 @@ on:
       - '[0-9]+.[0-9]+.[0-9]+-[A-Za-z0-9]+' # release branches
     paths:
       - 'distribution/**'
+      - 'web-console/**'
       - '**/pom.xml'
   pull_request:
     branches: