--- a/distribution/docker/Dockerfile
+++ b/distribution/docker/Dockerfile
@@ -23,6 +23,9 @@ ARG JDK_VERSION=17
 # This can be unset if the tarball was already built outside of Docker
 ARG BUILD_FROM_SOURCE="true"
 
+#===============================================================
+# web-console building stage
+#===============================================================
 # The platform is explicitly specified as x64 to build the Druid distribution.
 # This is because it's not able to build the distribution on arm64 due to dependency problem of web-console. See: https://github.com/apache/druid/issues/13012
 # Since only java jars are shipped in the final image, it's OK to build the distribution on x64.
@@ -37,23 +40,27 @@ COPY ./docs /src/docs
 WORKDIR /src/web-console
 
 # Re-declare the ARG to make it available in this stage
-ARG BUILD_FROM_SOURCE="true"
+ARG BUILD_FROM_SOURCE
 
+# Suppress a building error saying error:0308010C:digital envelope routines::unsupported
 ENV NODE_OPTIONS=--openssl-legacy-provider
 
-# The ls command below is just for debug helping us know which files are updated that docker cache are invalidated
+# The directory of output, during the package stage, files will be copied from this directory
+RUN mkdir /output
+
+# The ls command below is just for debug purpose helping us know which files are updated that docker cache are invalidated
 RUN if [ "$BUILD_FROM_SOURCE" = "true" ]; then \
   ls -l && \
   npm ci \
 ; fi
 RUN if [ "$BUILD_FROM_SOURCE" = "true" ]; then \
    ./script/build \
-   && ./script/cp-to ./target/resources \
+   && ./script/cp-to ./output \
 ; fi
 
-#
+#===============================================================
 # Declare a builder stage to build the Druid distribution
-#
+#===============================================================
 FROM maven:3.9.9-eclipse-temurin-${JDK_VERSION} as builder
 RUN apt-get -qq update \
     && apt-get -qq -y install --no-install-recommends python3 python3-yaml
@@ -62,7 +69,7 @@ COPY . /src
 WORKDIR /src
 
 # Copy the web-console resources from the web-console-builder to include it in this stage to package
-COPY --from=web-console-builder /src/web-console/target/resources /src/web-console/target/resources
+COPY --from=web-console-builder /output /src/web-console/target/resources
 
 # Re-declare the ARG to make it available in this stage
 ARG BUILD_FROM_SOURCE
@@ -80,11 +87,16 @@ RUN --mount=type=cache,target=/root/.m2 if [ "$BUILD_FROM_SOURCE" = "true" ]; th
       -T1C \
 ; fi
 
+
+# RUN ls ./distribution/target
 RUN DISTRIBUTION=$(ls ./distribution/target/apache-druid-*-bin.tar.gz) \
  && tar -zxf ${DISTRIBUTION} -C /opt \
  && mv /opt/apache-druid-* /opt/druid
 
-# The bulding stagings use images based on eclipse-temurin,
+#===============================================================
+# Final stage
+#===============================================================
+# Above bulding staging uses images based on eclipse-temurin,
 # however here we directly use alpine because eclipse-temurin-alpine for JRE17 does not ship a image with ARM64 architecture.
 FROM alpine:latest
 LABEL maintainer="Apache Druid Developers <dev@druid.apache.org>"