--- a/web-console/pom.xml
+++ b/web-console/pom.xml
@@ -34,6 +34,10 @@
   <properties>
     <resources.directory>${project.build.directory}/resources</resources.directory>
     <web.console.skip>false</web.console.skip>  <!-- this property is overidden in Travis CI to skip the javascript-related work -->
+
+    <!--
+    NOTE: When changing the following properties, the docker base image specified in distribution/docker/Dockerfile must be updated accordingly
+    -->
     <node.version>v20.9.0</node.version>
     <npm.version>10.9.0</npm.version>
   </properties>
@@ -150,6 +154,9 @@
         </executions>
       </plugin>
       <plugin>
+        <!--
+        The web.console.skip is NOT applied to this stage as this is required for web-console package when building docker image
+        -->
         <groupId>org.apache.maven.plugins</groupId>
         <artifactId>maven-resources-plugin</artifactId>
         <configuration>