@kgyrtkirk 2025-02-18
I was taking a look and was wondering about the following:
* I feel like the `BUILD_FROM_SOURCE` option is very wierd; why build these inside docker instead of packaging the realse into a docker image?
** is it possible with an M1 to build the dist build on the host and avoid building inside docker?
* I don't really like that the new docker build customizes the distribution build logic inside the Dockerfile - with the hazard of using differently versioned tools
* I think the `web-console` module doesn't correctly support incremental builds; so it gets rebuild every time; I think fixing that would also make these things painfull - using the docker cache implicitly adds a incremental build option...