@FrankChen021 2025-02-19
Building the distribution tarball outside docker is not a good idea, because we can't guarantee the host machine installs the correct version of jdk. So this would be only an option for local development.