@kgyrtkirk 2025-02-19
I don't think the installed jdk should matter - what could that alter? 
We are building with the `--release` option - so the runtime will be right.
Even the release guideline doesn't highlight that at all [here](https://github.com/apache/druid/blob/master/distribution/asf-release-process-guide.md#build-artifacts)
I believe this approach would only be used during development so I don't think it would matter that much.

So if you are able to build it then that should be the best - you will rely on maven to do the incremental build ; if you want to do a full rebuild ; you could `git clean -dfx`...

I would really like to know if you are able to build or not the project in a reasonable time on your M1 system with `BUILD_FROM_SOURCE=false` set?