@FrankChen021 2025-02-20
why doesn't the jdk matter if the BUILD_FROM_SOURCE is changed to false? Dockerfile strictly defines the building and running environment, why should we always build the image outside the docker while the host machine might have its own JDK release? 

Like for me, JDK8, JDK17, JDK21 are both installed, before JDK17 is well supported, I have to use `export JAVA_HOME` to specify using of JDK8. And now earlier JDK21 is also buggy to build druid, someone may also need to switch JDK before building. Turning this option into false by default introduces some extra work for some existing users.

`BUILD_FROM_SOURCE=false` definitely saves the building time(no need to prepare docker building context, no need to copy files into docker, use local .m2 cache, or skipping building of already built web-console by -Dweb.console.skip). This is based on a full understanding of how Druid is built. To build a official image or in any CICD pipeline, turning this to false should not be an option.

This option should be retired in the future as the web-console module can benefit from the docker cache now.