@kgyrtkirk 2025-02-20
I believe we see some things differently

> why doesn't the jdk matter if the BUILD_FROM_SOURCE is changed to false? 
I think its not taht important as long as you are using a decent one which is supported 

* jdk8 is not supported anymore - that's why we are able to use the `release` option
* jdk11 reached eol in 2024 and is not really supported anymore; it should probably be removed
* jdk17 should work well
* the jdk21 issue I know of is runtime specific; so I doubt the compiler would be affected - if its something else; please let me know

I think this is mostly a workaround which adds complexity because:
1. some hardware issues due to using Apple M1 is not supported for `node-sass`
2. web-console's incremental build is broken; so maven can't skip it correctly
3. web-console depends on `node-sass` (but doesn't use it AFAIK)
   * `node-sass` is [not maintained](https://sass-lang.com/blog/node-sass-is-end-of-life/)

I understand you can't fix `1` - but instead of fixing `2` or `3` you want to add complexity to the distribution build.
Having `2` is bad - fixing the incremental build of web-console would be beneficial in everyday developer work as well ; as it keeps popping up in builds when someone forgets to exclude that module from the build.
I don't know what would be the option to fix `3` ; but that would be better as well ...so less outdated stuff would be needed.

I believe if you want to build the distribution in Docker that's a different thing/task/etc; it should not be an integral part of building the Docker images of the project.
If that would be a way to build the dist build - I guess that would be probably a more reliable way to build the release as well - as doing so would ensure that the same classes would be packaged into both the binary distribution and into the docker images.

The current CI system uses github-actions which have its own ideas/ways to set things up - so Docker is not used there right now.