@FrankChen021 2025-02-21
> * jdk8 is not supported anymore - that's why we are able to use the `release` option
> * jdk11 reached eol in 2024 and is not really supported anymore; it should probably be removed
> * jdk17 should work well
> * the jdk21 issue I know of is runtime specific; so I doubt the compiler would be affected - if its something else; please let me know

Here're I think you're talking about the suitable JDK for Druid. However, we are not only working on Druid, but also other projects that uses different JDKs